// monolithic-system/src/modules/client-adm/usecase/find-client/find-client.usecase.dto.ts
export interface FindClientInputDto {
    id: string
}

export interface FindClientOutputDto {
    id: string
    name: string
    email: string
    document: string
    address: {
        street: string
        number: string
        complement: string
        city: string
        state: string
        zipCode: string
    }
    createdAt: Date
    updatedAt: Date
}